Removed guava completely
