+++ logbook-api/src/main/java/org/zalando/logbook/NullSafe.java
@@ -2,7 +2,7 @@ package org.zalando.logbook;
 
 /*
  * #%L
- * Logbook: Core
+ * Logbook: API
  * %%
  * Copyright (C) 2015 - 2016 Zalando SE
  * %%
