+++ logbook-spring-boot-starter/pom.xml
@@ -30,10 +30,6 @@
             <artifactId>logbook-core</artifactId>
         </dependency>
         <dependency>
-            <groupId>com.google.guava</groupId>
-            <artifactId>guava</artifactId>
-        </dependency>
-        <dependency>
             <groupId>org.zalando</groupId>
             <artifactId>logbook-httpclient</artifactId>
         </dependency>
