+++ logbook-test/pom.xml
@@ -33,10 +33,6 @@
             <groupId>org.zalando</groupId>
             <artifactId>logbook-api</artifactId>
         </dependency>
-        <dependency>
-            <groupId>com.google.guava</groupId>
-            <artifactId>guava</artifactId>
-        </dependency>
     </dependencies>
 
     <distributionManagement>
