+++ logbook-test/src/main/java/org/zalando/logbook/MockHttpRequest.java
@@ -29,7 +29,7 @@ import java.util.Collections;
 import java.util.List;
 import java.util.Map;
 
-import static com.google.common.base.MoreObjects.firstNonNull;
+import static org.zalando.logbook.NullSafe.firstNonNull;
 
 @Immutable
 public final class MockHttpRequest implements MockHttpMessage, HttpRequest {
