+++ logbook-test/src/main/java/org/zalando/logbook/MockRawHttpRequest.java
@@ -30,7 +30,7 @@ import java.util.Collections;
 import java.util.List;
 import java.util.Map;
 
-import static com.google.common.base.MoreObjects.firstNonNull;
+import static org.zalando.logbook.NullSafe.firstNonNull;
 
 @Immutable
 public final class MockRawHttpRequest implements MockHttpMessage, RawHttpRequest {
